requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.93.0
flower==1.2.0
gunicorn==20.1.0
h11==0.14.0
humanize==4.6.0
idna==3.4
kombu==5.2.4
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pydantic==1.10.6
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tornado==6.2
typing_extensions==4.5.0
uvicorn==0.21.1
vine==5.0.0
wcwidth==0.2.6
